Add tests for additional benefits page
